os/softhostid.c: commonize the _SOFT_HOSTID code such that we catch up with x86
This brings in:
  9e3e4df2a0 3359 hostid generation should be more "predictable"
  2df37efa12 3518 Incorrect hostid calculation

and then fixes them so the blacklist actually works.
